hw4/auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-asyncio
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
prometheus-fastapi-instrumentator
tenacity
python-logstash-async
